requirements.txt
pyubx2>=1.3.0
requests>=2.28.0
pyserial>=3.5
